fix cuda issue

+++ launch_benchmark.sh
@@ -12,9 +12,11 @@ function main {
     set_environment
 
     pip install -r ${workload_dir}/requirements.txt
-    pip uninstall -y numba llvmlite
-    conda install -c numba llvmdev -y
-    pip install git+https://github.com/numba/llvmlite.git
+    if [ "${device}" != "cuda" ];then
+        pip uninstall -y numba llvmlite
+        conda install -c numba llvmdev -y
+        pip install git+https://github.com/numba/llvmlite.git
+    fi
     #pip install -U numba
     pip install numba==0.48
     # pip install --no-deps torchvision -f https://download.pytorch.org/whl/torch_stable.html
